backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.0
pillow==10.0.1
opencv-python==4.8.1.78
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pydantic-settings==2.0.3 